登入攔截器
1.新增會員中心頁面，該頁面限定會員登入後才可以查看
2.登入時，將會員id存入session
3.新增類別interceptor，並配置好文件
4.若session中沒有會員id，強制彈回登入頁面
5.新增登出功能，即清除會員id的session

+++ Store/sql.sql
@@ -10,7 +10,7 @@ select * from user_table;
 
 desc user_table;
 
-delete from user_table;
+delete from user_table where user_id=28;
 
 insert into user_table(username, password) values('aaa', 'aaa');
